More flags
Using the flags package, use a lot more flags for variables. Not a fan
of having to default the real and imaginary midpoints to -100 just to
detect when to randomise - I guess really the randomised midpoint
should be outside of the program - i.e. this should be a library and
something valls it